% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_geoid.R
\name{label_geoid}
\alias{label_geoid}
\title{Label a US location data set with a GEOID column}
\usage{
label_geoid(data)
}
\arguments{
\item{data}{a data.frame with a GEOID column that is of class character}
}
\value{
data.frame like object with values labeled
}
\description{
Label a US location data set with a GEOID column with appropriate
state/county names
}
\examples{
label_geoid(state_total_df)

}
