% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_data.R
\name{project_data}
\alias{project_data}
\title{Project a long yearly data set forward holding last value observed constant}
\usage{
project_data(data, future_years)
}
\arguments{
\item{data}{a data.frame with a YEAR column which is long}

\item{future_years}{int vector years to project for}
}
\value{
data.frame like object with projected values
}
\description{
Project a long yearly data set forward holding last value
observed constant
}
\examples{
project_data(state_total_df, 2019)

}
